requirements.txt
asgiref==3.8.1
astroid==3.3.5
dill==0.3.9
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
Faker==30.3.0
isort==5.13.2
mccabe==0.7.0
pillow==10.4.0
platformdirs==4.3.6
PyJWT==1.7.1
pylint==3.3.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
